refactor: Выделить ConstructorPage из App
