+++ src/components/app/app.jsx
@@ -1,18 +1,14 @@
 import React, { useEffect } from 'react';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import '../../style/common.css';
 import styles from './app.module.css';
 import AppHeader from '../app-header/app-header';
-import BurgerIngredients from '../burger-ingredients/burger-ingredients';
-import BurgerConstructor from '../burger-constructor/burger-constructor';
 import '@ya.praktikum/react-developer-burger-ui-components';
 
 import { useDispatch, useSelector } from 'react-redux';
-import { DndProvider } from 'react-dnd';
-import { HTML5Backend } from 'react-dnd-html5-backend';
 import { fetchIngredients } from '../../services/slices/ingredientsSlice';
-import { selectIngredientsSlice } from '../../utils/selectors';
+import ConstructorPage from '../../pages/constructor';
 function App() {
-  const ingredients = useSelector(selectIngredientsSlice);
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -23,25 +19,13 @@ function App() {
     <div className={styles.app}>
       <AppHeader />
       <main className={`${styles.main}`}>
-        {(ingredients.isLoading || ingredients.error) && (
-          <p className={`text text_type_main-large text_color_inactive ${styles.message}`}>
-            {ingredients.isLoading
-              ? 'Загрузка данных'
-              : ingredients.error
-              ? 'Ошибка загрузки данных'
-              : ''}
-          </p>
-        )}
-        {!ingredients.isLoading && !ingredients.error && (
-          <>
-            <DndProvider backend={HTML5Backend}>
-              <BurgerIngredients />
-              <div className='ml-10 pt-25'>
-                <BurgerConstructor />
-              </div>
-            </DndProvider>
-          </>
-        )}
+        <Router>
+          <Switch>
+            <Route path={'/'} exact={true}>
+              <ConstructorPage />
+            </Route>
+          </Switch>
+        </Router>
       </main>
     </div>
   );
